backend/foodgram/api/views.py
from rest_framework import viewsets
from djoser.views import UserViewSet

from business_logic.models import Ingredient, Recipe, Tag
from users.models import User
from .serializers import (
    CustomUserSerializer, IngredientSerializer, RecipeSerializer, TagSerializer
)


class CustomUserViewSet(UserViewSet):
    queryset = User.objects.all()
    serializer_class = CustomUserSerializer


class IngredientViewSet(viewsets.ModelViewSet):
    queryset = Ingredient.objects.all()
    serializer_class = IngredientSerializer


class RecipeViewSet(viewsets.ModelViewSet):
    queryset = Recipe.objects.all()
    serializer_class = RecipeSerializer

    def perform_create(self, serializer):
        serializer.save(author=self.request.user)


class TagViewSet(viewsets.ModelViewSet):
    queryset = Tag.objects.all()
    serializer_class = TagSerializer
